Add testplan command
It can be useful to run multiple different builds in a sequence to test
that each of those builds runs through successfully.

Add command that takes a test plan YAML with multiple buildrun settings.

Add code to execute buildruns from a list of test plan steps.

Improve error message in case a buildrun fails to include all the logs
of all the containers of the respective taskrun pod.