Switch to wait polling style
The code used to use the client-go watcher to listen to all the events
that came in for a specific buildRun to eventually decide whether it
succeeded or failed. However, under certain high load conditions, some
of the events went missing. The buildRun was done, but the code did not
pick it up.

Change wait function to polling style.