Update check function to warn but not fail
There are environments, where the permissions do not allow to check for
system wide settings, like looking up cluster build strategies, or
listing nodes. In those environments, the tool should try to proceed
with a build as far as it can.

Update `CheckSystemAndConfig` function to not fail if the look ups fail,
but to gracefully go over it and print a warning to bring up the issue.