Fix wrong system check function
Due to the rewrite to allow more errors due to permissions, there were
bugs introduced because of the new logic and possible nil values.

Introduce specific checks for the error to improve the error paths.

Fix wrong error is nil check for nodes.

Explicitly set the clusterBuildStrategy to nil in error case.