Add namespace override for testplan
When the testplan is already provided and contains a namespace that is
not available in the current cluster, a manual change in the testplan
would be required. It would be easier if an override using a command
line flag could take precedence over the configuration file.

Add flag `namespace` to override the namespace defined in the testplan.