Rework calculations to fix misleasing output
The output summary mixed two things:
- The min and max were actually the exact buildrun times based on the
  sorted list of buildruns.
- The mean and median were newly created buildrun results with values
  that were calculated based on the input buildrun results.
This setup is confusing and not intuitive.

Change calculation so that both min and max are the smallest and biggest
values of all the values. With other words, all the times are now
calculated just like mean and median already were.

Drop rounding to seconds in calculation to have more precise data.

Remove repetitive code style by introducing list of values instead.