Fix issue with timeout on buildruns
With changes to the Shipwright internals, especially the removal of the Status
field the tool failed to correctly identify the completion of a buildrun.

Bump to latest available https://github.com/shipwright-io/build.

Update function names and interface names to match new names.

Rework polling and error function to work with conditions.